requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.3
redis[asyncio]==5.0.1
celery==5.3.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
